test: fix skipping of this repair test
Fix the usage of delay_dev so the test is properly executed
and not skiped. For major mirror slowdown use smalled region
sizes that are cause way more frequent commits so we can go
with significantly smaller delays.

Also check repair to work for failing mirror leg and mirror log.