persist: do not use pr_key config setting for sanlock VGs v2
Use only hostid-based PR keys for sanlock VGs.
This will be required for PR-based recovery with sanlock.

Changes how persist_start() is used to update an existing key.
Previously, the updated key was passed to persist_start as if
it was the local pr_key setting.  It's now passed as a separate
parameter.

This also fixes an incorrect warning in vgchange --persist check
when checking a hostid-based key while the lockspace is stopped.