persist: check key generation when writing or activating
When PR is used for fencing with a sanlock VG, the PR key
needs to include the correct sanlock generation number.
If it does not, then another host could remove the wrong
PR key for a failed host.  The PR key of a failed host is
calculated using the host_id and generation of the failed
host.

A host now ensures that its own sanlock generation matches
its own PR key generation each time it attempts to modify
or activate a VG.  Each lockd_vg() returns the current
generation, which is passed to persist_is_started_gen(),
which checks that the generation matches between the
PR key and the sanlock lockspace.  With this check, a
host is unable to modify or activate a VG unless it's
PR key is consistent with what other hosts expect.

The existing "vgchange --persist check" already detects
that the incorrect key is registered.

Running "vgchange --persist start" updates the registered
key with the correct generation.