lvmcache: fix warning about updating old metadata in vg_write
When vgsplit moves a PV from one VG to another, the check for
updating old VG metadata was incorrect and needs to be skipped.

The invalid check would produce an incorrect warning message:
WARNING: Updating old metadata to N on /dev/foo for VG bar.

The warning would only appear if the destination VG happened to
have a larger seqno than the source VG.