lvresize: fix xfs size checks
After lvextend with xfs_growfs on a mounted xfs file system,
libblkid FSLASTBLOCK continues to return the original value,
and does not reflect the grown size of the fs.

If that was followed by lvreduce back to the original LV size,
lvm saw the original fs size in FSLASTBLOCK, and allowed the LV
to shrink back to the original size, even though xfs had grown
and was now using the space.

Fix this by using the XFS_IOC_FSGEOMETRY ioctl to get the
correct size value for mounted xfs.